Consistently use int64 not uint64 when we have to work with time in micro or milli seconds. This is consistent with the behavior of the underlying classes.
Merge internal change: 38559304

Review URL: https://chromiumcodereview.appspot.com/11293220

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@167364 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/quic/quic_connection_helper_test.cc
@@ -209,7 +209,7 @@ TEST_F(QuicConnectionHelperTest, TimeoutAfterSend) {
   // This time, we should time out.
   EXPECT_CALL(visitor_, ConnectionClose(QUIC_CONNECTION_TIMED_OUT, false));
   runner_->RunNextTask();
-  EXPECT_EQ(uint64(kDefaultTimeoutUs + 5000), clock_.Now().ToMicroseconds());
+  EXPECT_EQ(kDefaultTimeoutUs + 5000, clock_.Now().ToMicroseconds());
   EXPECT_FALSE(connection_.connected());
 }
